feat(EMS-607): update task list to render single or multiple policy type specific fields/task requirements
